backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-chess>=1.999
pydantic>=2.5.0
